アニメレビュー関連のエンドポイントを分割しました
