+++ node_server/api/api.js
@@ -17,6 +17,7 @@ const animeModel = {
 
 const getMinogashi = require('./getMinogashiAnime');
 const getShoboiAPI = require('./getShoboiAPI');
+const getReview = require('./getReview');
 
 if(animeModel.review == null){
   return;
@@ -38,6 +39,12 @@ getMinogashi.getAll(router, animeModel);
 
 // アニメの見逃し情報を取得する
 getMinogashi.getWeek(router, animeModel);
+
+// レビューを1件返す
+getReview.getOne(router, animeModel);
+
+// 全件のレビューを返す
+getReview.getAll(router, animeModel);
 
 // アニメの全話数情報を取得
 router.get("/getAnimeStoryList/:tid", (req, res)=>{
@@ -134,62 +141,6 @@ router.post("/setAnimeAbout", (req, res)=>{
       status: 'ng',
       date: moment(),
       e: e
-    };
-    res.send(res_body);
-  });
-});
-
-// レビューを1件返す
-router.get("/getAnimeReview/:tid", (req, res)=>{
-  res.header('Content-Type', 'application/json');
-  let animePromises = [];
-  animePromises.push( animeModel.review.getAnimeReview(req.params.tid) );
-  animePromises.push( animeModel.about.getAnime(req.params.tid) );
-  Promise.all(animePromises)
-  .then((db_data)=>{
-    // レビューがない場合には新規作成する
-    if(db_data[0].length == 0){
-      return animeModel.review.initReview(req.params.tid)
-      .then((init_res)=>{
-        let res_body = {
-          status: 'ok make new review',
-          body: null
-        };
-        res.send(res_body);
-      });
-    }
-    else{
-      let res_body = {
-        status: 'ok',
-        body: Object.assign(db_data[0][0].dataValues, db_data[1][0].dataValues)
-      };
-      res.send(res_body);
-    }
-  })
-});
-
-// 全件のレビューを返す
-router.get("/getAllAnimeReview", (req, res)=>{
-  let animePromises = [];
-  animePromises.push( animeModel.review.getAllAnimeReview() );
-  animePromises.push( animeModel.about.getAllAnime() );
-  Promise.all(animePromises)
-  .then((db_datas)=>{
-    let animeList = [];
-    let animeReviewList = db_datas[0];
-    let animeAboutList = db_datas[1];
-    for(let i=0; i<animeAboutList.length; i++){
-      for(let j=0; j<animeReviewList.length; j++){
-        if(animeAboutList[i].tid == animeReviewList[j].tid){
-          animeList.push( Object.assign(animeAboutList[i].dataValues, animeReviewList[j].dataValues) );
-          break;
-        }
-      }
-    }
-    res.header('Content-Type', 'application/json');
-    let res_body = {
-      status: 'ok',
-      body: animeList
     };
     res.send(res_body);
   });
